package structure cleanup
