+++ .pre-commit-config.yaml
@@ -27,4 +27,4 @@ repos:
     hooks:
     - id: isort
       name: isort
-      args: [--line-length=130]
+      args: [--line-length=120]
