+++ gpt2/midi_datasets/AugmentedDataset/AugmentedDatasetConfig.py
